refactor(observability): move facade behind ObservabilityStoragePort (Slice 3 Stage B)
Completes the observability root-crate extraction. The facade —
ObservabilityService (+ ingestion/query param-result structs),
ObservabilityIngester, and ObservabilityQueryEngine — moves from the root into
proximadb-observability-engine, now holding Arc<dyn ObservabilityStoragePort>
(dependency inversion). The root WAL-backed ObservabilityStorage impls the port
(Stage A); composition roots inject it, auto-coerced at the call sites.

ObservabilityStoragePort expanded 8 -> 11 methods (+query_logs, query_metrics,
query_traces_by_service) to cover the full surface the facade actually calls.

The root-local `impl ObservabilityStorageOperations for ObservabilityService`
stays in the root crate (orphan rule: the trait is local to root). It now reaches
the service through two new pub accessors -- storage() and namespace_details() --
instead of private fields.

The 7 query-facade inline tests (which construct the root ObservabilityStorage)
move to src/observability/query_tests.rs; parse_query / parse_severity /
ParsedQuery are now pub so the root tests exercise them via the crate re-export.

Left in root (out of scope): ingestion format adapters (OTLP/Syslog/CEF/... --
axum/tonic/tower ingress servers, network-layer) and graph_linking (orphan/dead
code, never declared in a module tree).

Net ~2.5K LOC out of the root crate. Storage up-edge ratchet unchanged at 197.
