+++ crates/modalities/proximadb-observability-engine/Cargo.toml
@@ -52,6 +52,8 @@ uuid = { version = "1.0", features = ["v4", "serde"] }
 # Alerting notification channels (email + webhook) — root-only deps, same specs.
 lettre = "0.11"
 reqwest = { version = "0.11", features = ["json"] }
+# Ingester worker sizing — root-only dep, same spec.
+num_cpus = "1.16"
 # Foundation types — the pure observability modules are foundation-clean
 # (they name data_model/kernel/records/tenant types, not proto messages and
 # not root storage/query/services).
